Don't explicitly enable CFI in libvpx.
Previously there was no other sanitization, so in order for the blacklist
to be picked up CFI had to be enabled. With integer_overflow enabled, we
no longer need it enabled explicitly in the build file. Since libvpx is
a static library, both CFI and non-CFI variants will be built and linked
accordingly.

If libvpx is used as a shared library in the future, we'll need to
explicitly re-enable CFI in the build file.

Bug: 63926619
Test: Library builds.
Change-Id: I6806eba2659cb047fc5e58de122f24d905bcd06c