libvpx: cherry pick compare_tile_buffers fix
vp9: address integer sanitizer warning

Comparing the size values with subtraction requires casting. Sort in
descending order.

(a < b) - (a > b)
If a is greater, this is 0 - 1 = -1
If the  values are equal, this is 0 - 0 = 0
If b is greater, this is 1 - 0 = 1

cherry picked from libvpx:
a532c243bb2b5bb0fd0ef295eb019518cc532ca5

Test: passes integer sanitizer builds
Change-Id: Idc4f34cfa05d672f46eed774bcd8e75abf8a157b